(ref): modularizacion metodos auxiliares, POM y Tests
