+++ helpers/util_helpers.py
@@ -1,0 +1,32 @@
+# Funciones Útiles
+# no modificar
+def retrieve_phone_code(driver) -> str:
+    """Este código devuelve un número de confirmación de teléfono y lo devuelve como un string.
+    Utilízalo cuando la aplicación espere el código de confirmación para pasarlo a tus pruebas.
+    El código de confirmación del teléfono solo se puede obtener después de haberlo solicitado en la aplicación."""
+
+    import json
+    import time
+    from selenium.common import WebDriverException
+    code = None
+    for i in range(10):
+        try:
+            logs = [log["message"] for log in driver.get_log('performance') if log.get("message")
+                    and 'api/v1/number?number' in log.get("message")]
+            for log in reversed(logs):
+                message_data = json.loads(log)["message"]
+                body = driver.execute_cdp_cmd('Network.getResponseBody',
+                                              {'requestId': message_data["params"]["requestId"]})
+                code = ''.join([x for x in body['body'] if x.isdigit()])
+        except WebDriverException:
+            time.sleep(1)
+            continue
+        if not code:
+            raise Exception("No se encontró el código de confirmación del teléfono.\n"
+                            "Utiliza 'retrieve_phone_code' solo después de haber solicitado el código en tu aplicación.")
+        return code
+
+def parse_to_seconds(time_str):
+    minutes, seconds = time_str.split(":")
+    total_seconds = int(seconds) + int(minutes) * 60
+    return total_seconds
