+++ main.py
@@ -1,296 +1,9 @@
-import time
-
 import data
 from selenium import webdriver
-from selenium.webdriver import Keys
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.common.exceptions import TimeoutException
 
-
-# Funciones Útiles
-# no modificar
-def retrieve_phone_code(driver) -> str:
-    """Este código devuelve un número de confirmación de teléfono y lo devuelve como un string.
-    Utilízalo cuando la aplicación espere el código de confirmación para pasarlo a tus pruebas.
-    El código de confirmación del teléfono solo se puede obtener después de haberlo solicitado en la aplicación."""
-
-    import json
-    import time
-    from selenium.common import WebDriverException
-    code = None
-    for i in range(10):
-        try:
-            logs = [log["message"] for log in driver.get_log('performance') if log.get("message")
-                    and 'api/v1/number?number' in log.get("message")]
-            for log in reversed(logs):
-                message_data = json.loads(log)["message"]
-                body = driver.execute_cdp_cmd('Network.getResponseBody',
-                                              {'requestId': message_data["params"]["requestId"]})
-                code = ''.join([x for x in body['body'] if x.isdigit()])
-        except WebDriverException:
-            time.sleep(1)
-            continue
-        if not code:
-            raise Exception("No se encontró el código de confirmación del teléfono.\n"
-                            "Utiliza 'retrieve_phone_code' solo después de haber solicitado el código en tu aplicación.")
-        return code
-
-def parse_to_seconds(time_str):
-    minutes, seconds = time_str.split(":")
-    total_seconds = int(seconds) + int(minutes) * 60
-    return total_seconds
-
-
-class UrbanRoutesPage:
-    # Localizadores de elementos en la página mediante diferentes estrategias
-    # Localizadores del formulario para establecer ruta
-    from_field_locator = (By.ID, 'from')
-    to_field_locator = (By.ID, 'to')
-    call_a_taxi_btn_locator = (By.XPATH, '//button[@type="button" and text()="Pedir un taxi"]')
-
-    # Localizadores de las tarifas
-    comfort_tariff_card_img_locator = (By.XPATH, '//div[@class="tcard-icon"]/img[@alt="Comfort"]')
-    setted_tariff_card_text_locator = (By.XPATH, '//div[@class="tcard active"]//div[@class="tcard-icon"]/following-sibling::div[@class="tcard-title"]')
-
-    # Localizadores del campo para agregar numero telefonico
-    phone_number_btn_locator = (By.XPATH, '//div[contains(@class,"np-button")]/div[@class="np-text"]')
-    phone_number_input_field_on_modal_locator = (By.XPATH, '//div[@class="input-container"]/input[@id="phone" and @placeholder="+1 xxx xxx xx xx"]')
-    phone_number_submit_btn_locator = (By.XPATH, '//button[@type="submit" and text()="Siguiente"]')
-    phone_code_field_locator = (By.XPATH, '//div[@class="input-container"]/input[@id="code" and @placeholder="xxxx"]')
-    phone_code_submit_btn_locator = (By.XPATH, '//button[@type="submit" and text()="Confirmar"]')
-
-    # Localizadores del campo para agregar metodo de pago
-    payment_method_btn_locator = (By.XPATH, '//div[contains(@class, "pp-button")]/div[@class="pp-text"]')
-    payment_method_setted_on_payment_method_btn = (By.XPATH, '//div[contains(@class, "pp-value")]/div[@class="pp-value-text"]')
-    add_card_btn_locator = (By.XPATH, '//div[@class="pp-title" and text()="Agregar tarjeta"]')
-    card_number_field_locator = (By.XPATH, '//div[@class="card-number-input"]/input[@id="number" and @placeholder="1234 4321 1408"]')
-    card_code_field_locator = (By.XPATH, '//div[@class="card-code-input"]/input[@id="code" and @placeholder="12"]')
-    new_card_submit_btn_locator = (By.XPATH, '//button[@type="submit" and text()="Agregar"]')
-    close_add_card_modal_btn_locator = (By.XPATH, '//div[@class="head" and text()="Método de pago"]/preceding-sibling::button[contains(@class, "close-button")]')
-
-    # Localizadores del campo para agregar un comentario para el conductor
-    message_for_the_driver_field_locator = (By.XPATH, '//input[@id="comment" and @placeholder="Traiga un aperitivo"]')
-
-    # Localizadores de los checkbox y los contadores para pedir mantas, panuelos y helados
-    blanket_n_handkerchiefs_slider_locator = (By.XPATH, '//div[@class="r-sw-label" and text()="Manta y pañuelos"]/following-sibling::div[@class="r-sw"]//input[@type="checkbox"]/following-sibling::span')
-    icecream_counter_field_locator = (By.XPATH, '//div[@class="r-counter-container"]/div[text()="Helado"]/following-sibling::div[@class="r-counter"]//div[@class="counter-value"]')
-    icecream_increase_counter_btn_locator = (By.XPATH, '//div[@class="r-counter-container"]/div[text()="Helado"]/following-sibling::div[@class="r-counter"]//div[@class="counter-plus" and text()="+"]')
-
-    # Localizadores del boton que inicia la busqueda de un conductor
-    call_the_vehicle_btn_locator = (By.XPATH, '//button[@type="button"]/span[text()="Pedir un taxi"]')
-
-    # Localizadores del modal que contiene informacion de la busqueda del conductor
-    searching_vehicle_modal_locator = (By.XPATH, '//div[@class="order-body" and .//div[@class="order-header-title" and text()="Buscar automóvil"]]')
-
-    # Localizadores para la informacion del conductor
-    driver_arrival_time_label_locator = (By.XPATH, '//div[@class="order-header-title" and text()="Buscar automóvil"]/following-sibling::div[@class="order-header-time"]')
-    driver_rating_field_locator = (By.XPATH, '//div[@class="order-buttons"]/div[@class="order-btn-group"]//div[@class="order-btn-rating"]')
-    driver_img_field_locator = (By.XPATH, '//div[@class="order-buttons"]/div[@class="order-btn-group"]//div[@class="order-btn-rating"]/following-sibling::img')
-    driver_name_field_locator = (By.XPATH, '//div[@class="order-buttons"]/div[@class="order-btn-group"]/div[@class="order-button" and ./div[@class="order-btn-rating"]]/following-sibling::div')
-
-    # Constructor de la clase, inicializa el controlador, y establace los tiempos de espera para recuperar elementos de la pagina y para inspeccion visual
-    def __init__(self, driver, search_element_timeout=5, visual_review_timeout=3):
-        self.driver = driver
-        self.wait = WebDriverWait(self.driver, search_element_timeout)
-        self.VISUAL_REVIEW_TIMEOUT = visual_review_timeout
-
-    # Métodos para interactuar con los elementos de la página
-    # Métodos para interactuar con el formulario para establecer la ruta (Desde y Hasta)
-    def set_from(self, from_address):
-        self.__set_element_text(self.from_field_locator, from_address)
-    def set_to(self, to_address):
-        self.__set_element_text(self.to_field_locator, to_address)
-    def get_from(self):
-        return  self.__get_element_text(self.from_field_locator)
-    def get_to(self):
-        return  self.__get_element_text(self.to_field_locator)
-    def set_route(self, address_from, address_to):
-        self.set_from(address_from)
-        self.set_to(address_to)
-    def click_on_call_a_taxi_btn(self):
-        self.__click_on_element(self.call_a_taxi_btn_locator)
-
-    # Métodos para intecarctuar con las tarifas
-    def click_on_comfort_tariff_card(self):
-        self.__click_on_element(self.comfort_tariff_card_img_locator)
-    def get_selected_tariff(self):
-        return self.__get_element_text(self.setted_tariff_card_text_locator)
-
-
-    # Métodos para agregar el numero de telefono
-    def click_on_phone_number_btn(self):
-        self.scroll_into_and_click_on_element(self.phone_number_btn_locator)
-    def set_phone_number(self, phone_number):
-        self.__set_element_text(self.phone_number_input_field_on_modal_locator, phone_number)
-    def get_phone_number_from_field_on_modal(self):
-        return self.__get_element_text(self.phone_number_input_field_on_modal_locator)
-    def get_phone_number_from_btn(self):
-        return self.__get_element_text(self.phone_number_btn_locator)
-    def click_on_submit_phone_number_btn(self):
-        self.__click_on_element(self.phone_number_submit_btn_locator)
-    def click_on_submit_phone_code_btn(self):
-        self.__click_on_element(self.phone_code_submit_btn_locator)
-    def set_phone_code(self, phone_code):
-        self.__set_element_text(self.phone_code_field_locator, phone_code)
-
-    # Métodos para interactuar con los elementos de la tarjeta
-    def click_on_payment_method_btn(self):
-        self.__click_on_element(self.payment_method_btn_locator)
-    def get_type_of_setted_payment_method_from_payment_method_btn(self):
-            return self.__get_element_text(self.payment_method_setted_on_payment_method_btn)
-    def click_on_add_card_btn(self):
-        self.__click_on_element(self.add_card_btn_locator)
-    def set_card_number(self, card_number):
-        self.__set_element_text(self.card_number_field_locator, card_number)
-    def get_card_number(self):
-        return self.__get_element_text(self.card_number_field_locator)
-    def set_card_code(self, card_code):
-        self.__set_element_text(self.card_code_field_locator, card_code)
-    def get_card_code(self):
-        return self.__get_element_text(self.card_code_field_locator)
-    def click_on_submit_new_card_btn(self, send_tab_to_remove_focus=False):
-        if send_tab_to_remove_focus:
-            self.__set_element_text(self.card_code_field_locator, Keys.TAB)
-        self.__click_on_element(self.new_card_submit_btn_locator)
-    def click_on_close_add_card_modal_btn(self):
-        self.__click_on_element(self.close_add_card_modal_btn_locator)
-    def add_new_card(self, card_number, card_code):
-        self.click_on_payment_method_btn()
-        self.click_on_add_card_btn()
-        self.set_card_number(card_number)
-        self.set_card_code(card_code)
-        self.click_on_submit_new_card_btn(send_tab_to_remove_focus=True)
-        self.click_on_close_add_card_modal_btn()
-
-    # Métodos para agregar un comentario para el conductor
-    def set_message_for_the_driver(self, comment):
-        self.__set_element_text(self.message_for_the_driver_field_locator, comment)
-    def get_message_for_the_driver(self):
-        return self.__get_element_text(self.message_for_the_driver_field_locator)
-
-    # Métodos para interactuar con el checkbox de manta y panuelos, y el contador de helados
-    def click_on_checkbox_blanket_and_handkerchiefs(self):
-        self.scroll_into_and_click_on_element(self.blanket_n_handkerchiefs_slider_locator)
-    def check_if_checkbox_blanket_and_handkerchiefs_is_checked(self):
-        # es necesario porque selenium no lo encuentra ya que sus dimensiones son 0x0
-        checkbox = self.driver.execute_script("""
-                                                const expectedLabelText = "Manta y pañuelos"
-                                                const containers = document.querySelectorAll('div.r-sw-container');
-                                                let checkbox = null;
-                                                
-                                                containers.forEach( (container) => {
-                                                    const label = container.querySelector('.r-sw-label');
-                                                    const currentlabelText = label.textContent;
-                                                    if (expectedLabelText === currentlabelText) checkbox = container.querySelector('.switch-input');
-                                                });
-                                                return checkbox;
-                                              """)
-
-        return checkbox.get_property('checked')
-    def get_icecream_counter_number(self):
-        return self.__get_element_text(self.icecream_counter_field_locator)
-    def clicks_on_icecream_increase_counter_btn(self, times=1):
-        self.__click_on_element_multiple_times(self.icecream_increase_counter_btn_locator, times)
-
-    # Métodos para iniciar la busqueda de un conductor
-    def click_on_call_the_vehicle_btn(self):
-        self.__click_on_element(self.call_the_vehicle_btn_locator)
-
-    # Métodos para comprobar que el modal de busqueda de vehiculo aparece
-    def check_if_appears_searching_vehicle_modal(self, search_timeout=5, minimum_visibility_time=1):
-        # El minimum_element_visibility_time es para comprobar que el elemento se muestre en pantalla como minimo x segundos,
-        # ya que hay casos en los que el elemento aparece, pero desaparece inmediatamente,
-        # ocasionando que selenium lo detecte y lo marque como que si esta presente
-        # cuando en realidad puede que lo querramos presente mas tiempo
-        return self.__check_if_appears_element(self.searching_vehicle_modal_locator, search_timeout, minimum_visibility_time)
-
-    # Métodos para obtener la informacion del conductor encontrado
-    def get_driver_rating(self):
-        return self.__get_element_text(self.driver_rating_field_locator)
-    def get_driver_img_url(self):
-        return self.wait.until(EC.visibility_of_element_located(self.driver_img_field_locator)).get_attribute('src')
-    def get_driver_name(self):
-        return self.__get_element_text(self.driver_name_field_locator)
-
-
-    # Metodos auxiliares para interactuar con los elementos de la pagina
-    def __click_on_element(self, element_locator):
-        self.wait.until(EC.element_to_be_clickable(element_locator)).click()
-        self.confirm_and_trigger_visual_review_timeout()
-
-    def __scroll_into_element(self, element_locator):
-        element = self.wait.until(EC.visibility_of_element_located(element_locator))
-        self.driver.execute_script("arguments[0].scrollIntoView();", element)
-        self.confirm_and_trigger_visual_review_timeout()
-
-
-    def __set_element_text(self, element_locator, text):
-        self.wait.until(EC.visibility_of_element_located(element_locator)).send_keys(text)
-        self.confirm_and_trigger_visual_review_timeout()
-    def __get_element_text(self, element_locator):
-        element = self.wait.until(EC.visibility_of_element_located(element_locator))
-
-        text = element.text
-        if text:
-            return text
-
-        text = element.get_property('value')
-
-        if text:
-            return text
-
-        return ""
-    def __click_on_element_multiple_times(self, locator, times=1):
-        element = self.wait.until(EC.element_to_be_clickable(locator))
-        for _ in range(times):
-            element.click()
-
-        self.confirm_and_trigger_visual_review_timeout()
-    def __check_if_appears_element(self, element_locator, search_timeout=5, minimum_element_visibility_time=2):
-        # El minimum_element_visibility_time es para comprobar que el elemento se muestre en pantalla como minimo x segundos,
-        # ya que hay casos en los que el elemento aparece, pero desaparece inmediatamente,
-        # ocasionando que selenium lo detecte y lo marque como que si esta presente
-        # cuando en realidad puede que lo querramos presente mas tiempo
-
-        # si no requerimos comprobar que el elemento esta presente como minimo x segundos
-        if minimum_element_visibility_time == 0:
-            try:
-                WebDriverWait(self.driver, search_timeout).until(EC.visibility_of_element_located(element_locator))
-            except TimeoutException:
-                return False
-            return True
-
-        initial_time = time.time()
-
-        # si requerimos comprobar que el elemento este presente como minimo x cantidad de segundos
-        try:
-            # comprobar permanencia del elemento localizandolo
-            while time.time() - initial_time < minimum_element_visibility_time:
-                WebDriverWait(self.driver, search_timeout).until(EC.visibility_of_element_located(element_locator))
-        except TimeoutException:
-            return False
-        return True
-
-    def scroll_into_and_click_on_element(self, element_locator):
-        self.__scroll_into_element(element_locator)
-        self.__click_on_element(element_locator)
-    def wait_for_visual_review(self):
-        time.sleep(self.VISUAL_REVIEW_TIMEOUT)
-    def confirm_and_trigger_visual_review_timeout(self):
-        if self.VISUAL_REVIEW_TIMEOUT > 0:
-            self.wait_for_visual_review()
-    def wait_for_driver_info(self, additional_secs=1):
-        time_str = self.__get_element_text(self.driver_arrival_time_label_locator)
-        arrival_time_in_seconds = parse_to_seconds(time_str)
-        total_timeout = arrival_time_in_seconds + additional_secs
-        WebDriverWait(self.driver, total_timeout).until(EC.visibility_of_element_located(self.driver_name_field_locator))
-    def clear_local_storage(self):
-        self.driver.execute_script("window.localStorage.clear();")
-
-
-
+from helpers.pom_helpers import clear_local_storage
+from urban_routes_page import UrbanRoutesPage
+from helpers.util_helpers import retrieve_phone_code
 
 class TestUrbanRoutes:
 
@@ -334,7 +47,7 @@ class TestUrbanRoutes:
                                       visual_review_timeout=visual_timeout)
 
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
@@ -363,7 +76,7 @@ class TestUrbanRoutes:
                                       search_element_timeout=search_timeout,
                                       visual_review_timeout=visual_timeout)
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
@@ -417,7 +130,7 @@ class TestUrbanRoutes:
                                       visual_review_timeout=visual_timeout)
 
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
@@ -483,7 +196,7 @@ class TestUrbanRoutes:
                                       visual_review_timeout=visual_timeout)
 
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
@@ -511,7 +224,7 @@ class TestUrbanRoutes:
                                       search_element_timeout=search_timeout,
                                       visual_review_timeout=visual_timeout)
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
@@ -543,7 +256,7 @@ class TestUrbanRoutes:
                                       search_element_timeout=search_timeout,
                                       visual_review_timeout=visual_timeout)
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
@@ -578,7 +291,7 @@ class TestUrbanRoutes:
                                       visual_review_timeout=visual_timeout)
 
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
@@ -631,7 +344,7 @@ class TestUrbanRoutes:
                                       search_element_timeout=search_timeout,
                                       visual_review_timeout=visual_timeout)
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
@@ -700,7 +413,7 @@ class TestUrbanRoutes:
                                       visual_review_timeout=visual_timeout)
 
         # Limpia el almacenamiento local
-        routes_page.clear_local_storage()
+        clear_local_storage(routes_page.driver)
 
         # Establece la ruta de la prueba, con la dirección de origen y destino proporcionadas en los datos.
         routes_page.set_route(data.address_from, data.address_to)
